make folders generic without glassfish version.
