+++ glassfish-common/pom.xml
@@ -1,13 +1,13 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <parent>
-    <artifactId>arquillian-parent-glassfish6</artifactId>
+    <artifactId>arquillian-container-glassfish-jakarta</artifactId>
     <groupId>org.jboss.arquillian.container</groupId>
     <version>1.0.0.Final-SNAPSHOT</version>
     <relativePath>..</relativePath>
   </parent>
-  <artifactId>arquillian-glassfish6-common</artifactId>
-  <name>Arquillian Container GlassFish 6+ Common Services</name>
+  <artifactId>arquillian-glassfish-common</artifactId>
+  <name>Arquillian Container GlassFish 7+ Common Services</name>
 
   <dependencies>
