--- glassfish-managed-6/src/main/resources/META-INF/services/org.jboss.arquillian.core.spi.LoadableExtension
@@ -1,1 +1,0 @@
-org.jboss.arquillian.container.glassfish.managed_6.GlassFishManagedContainerExtension
