+++ glassfish-managed/src/main/resources/META-INF/services/org.jboss.arquillian.core.spi.LoadableExtension
@@ -1,0 +1,1 @@
+org.jboss.arquillian.container.glassfish.managed.GlassFishManagedContainerExtension
