+++ glassfish-managed/src/test/resources/application.xml
