+++ glassfish-managed/src/test/resources/arquillian.xml
