+++ glassfish-remote/src/test/java/org/jboss/arquillian/container/glassfish/remote/CDIJarTestCase.java
