+++ glassfish-remote/src/test/java/org/jboss/arquillian/container/glassfish/remote/SimpleBean.java
