+++ glassfish-remote/src/test/resources/application.xml
