+++ glassfish-remote/src/test/resources/arquillian.xml
